FILE: moves getOptions
