+++ libs/spirit-islander/util/src/lib/functions/create-game-setup.ts
@@ -3,7 +3,7 @@ import type { Config } from '../app/config.interface';
 import type { GameSetup } from '../app/game-setup.interface';
 import type { Difficulty } from '../game/difficulty';
 import { getValidCombos } from './get-valid-combos';
-import { getOptionsByName } from './get-options';
+import { getOptionsByName } from '../game/get-options';
 import { getDifficulty } from '../game/get-difficulty';
 import { selectBoards } from './select-boards';
 import { selectRandom } from './internal/select-random';
